avoid symmetric comparisons
This avoids feeding some extra hypotheses to linarith,
by mapping over the upper diagonal of a product of lists
instead of the full product.
Ultimately it doesn't matter, as linarith will filter out the redundant comparisons.
But this is better for performance,
and if linarith changes to use lists instead of rb_sets,
these will no longer be filtered.