feat(tactic/interactive_expr): always select all arguments (#3384)
If you hover over `id id 0` in the widget (or really any function with more than one argument), then it is possible to select the partial application `id id`.  The popup will then only show the function `id` and the argument `id`, but not the second argument `0`.

This PR changes this behavior so that you can't select partial applications and always see all argument in the popup.