feat(geometry/euclidean): orthogonal projection
Define orthogonal projection onto an affine subspace of a Euclidean
affine space, and prove some basic lemmas about it.

---

This depends on #3563 (orthogonal projection in real inner product
spaces).  Both this PR and #3615 (affine space lemmas) are in turn
dependencies of defining `circumcenter` and `circumradius` for a
`simplex` in a Euclidean affine space.  I have further lemmas about
orthogonal projection in relation to points equidistant from a set of
points, but it seems natural for those to go in the PR for
`circumcenter` and `circumradius` rather than the present PR.