feat(tactic/linarith): support case splits in preprocessing (including `ne` hypotheses) (#3786)
This adds an API for `linarith` preprocessors to branch. Each branch results in a separate call to `linarith`, so this can be exponentially bad. Use responsibly.

Given that the functionality is there, and I needed a way to test it, I've added a feature that people have been begging for that I've resisted: you can now call `linarith {split_ne := tt}` to handle `a ≠ b` hypotheses. Again: 2^n `linarith` calls for `n` disequalities in your context.



Co-authored-by: Rob Lewis <rob.y.lewis@gmail.com>
Co-authored-by: Mario Carneiro <di.gama@gmail.com>