Idea: do some simple line-break logic in lean
I ackowledge that it doesn't currently look very good, but the basic
idea is there. The idea is just that lean keeps track of string lengths
and then inserts some linebreaks when it overruns the 'max_width'
hardcoded constant. In the future it would get this from the extension
but for now if it's just set to something sensible I think that would be
fine. I've set the max-width to 20 characters to demonstrate.

Still some work to do:
- adding the 'div' breaks the highlighting logic.
- currently looks terrible, but the principle is there.