fix(algebra/field, ring_theory/ideal/basic): fix the issue that a predicate was given under square brackets
Seems like there is a problem when giving as an argument under square
brackets a predicate. I did not get this error on my device and I did
try leanproject build. The issue has been fixed along with a few other
minor changes to the features of my previous commit.